Updated trainingrun.sh
It now makes the user 'ls' once they are in the directory to 'look around' more.

There is another bug in the program. It does show '0 completed out of 3' once the user 'ls' to look at all the directories. I will try to see what I can do to fix that. I have messed with it, but nothing I've tried has worked yet.