Fixed 'completion' error when listing results.
The completion out of 3 no longer appears when the user commands 'ls' to list the results. It now works when the user enters whatever directory, and counts the number of completion out of 3 accordingly, as well as prompts the user to use 'ls' when they enter the directory, and prints the description.