ponygo file for Section Two & Developer tools
Still needs some edits and reformating.
Once we test it in the game, I can go back and reformat things if need be.
Also going back to add other minigames/exercises.

I also added the resource file for intro to bash scripting/bash that will be helpful for novice developers.